Change building instructions to favor official distribution
The original brief building instructions started with a git
clone, which works fine but encourages new users to grab the
latest development snapshot. It is probably better for new
users to use a release. Change the instructions to show
using a tarball. Even if the version changes, the general
idea should be captured.