Exposed the edges intersected by Marching Cubes.
Exposed the indices of the interpolated edges by the Marching Cubes worklet to the Marching Cubes filter. This way users have direct access to them to do further processing of the isosurface mesh. For example labeling the triangles based on which connected component they belong to. This is useful because inferring the edge directly from the vertices of the triangles of the MC output data set can be amgious due to numerical error.